requirements.txt
pymongo
fastapi
uvicorn
pydantic
passlib
python-dotenv
pydantic[email]
bcrypt
fastapi-mail
pandas
numpy
scipy
scikit-learn
matplotlib